feat: Add remaining tests for Matrix class
