+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,11 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\src",
+    "\\src\\csharp",
+    "\\src\\csharp\\Matrix",
+    "\\src\\csharp\\Tests"
+  ],
+  "SelectedNode": "\\src\\csharp\\Prelude.sln",
+  "PreviewInSolutionExplorer": false
+}
