+++ src/csharp/Prelude.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Matrix", "Matrix\Matrix.csproj", "{3FE1D08C-C90C-433E-B28A-9A1CBEAF08D0}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{D2749768-4C7A-477A-80E9-5E150E9058A5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +32,17 @@ Global
 		{3FE1D08C-C90C-433E-B28A-9A1CBEAF08D0}.Release|x64.Build.0 = Release|Any CPU
 		{3FE1D08C-C90C-433E-B28A-9A1CBEAF08D0}.Release|x86.ActiveCfg = Release|Any CPU
 		{3FE1D08C-C90C-433E-B28A-9A1CBEAF08D0}.Release|x86.Build.0 = Release|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Debug|x64.Build.0 = Debug|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Debug|x86.Build.0 = Debug|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Release|x64.ActiveCfg = Release|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Release|x64.Build.0 = Release|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Release|x86.ActiveCfg = Release|Any CPU
+		{D2749768-4C7A-477A-80E9-5E150E9058A5}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
